Atualizando backend, adicionando tipos a cada user, adicionando mais endpoints, criando page de admin, atualizando pagina de login

+++ backend/src/routes/mongoRoutes.js
@@ -13,6 +13,28 @@ router.get("/listTarefs", async (req, res) => {
   }
 });
 
+router.get("/listTarefs/user/:userId", async (req, res) => {
+  try {
+    const listers = await listTarefs.findById({ userId: req.params.userId });
+  } catch (err) {
+    return res.status(500).json({ error: err.message });
+  }
+});
+
+router.get("/listTarefs/:id", async (req, res) => {
+  const { id } = req.params;
+  try {
+    const list = await listTarefs.findById(id);
+    if (!list) {
+      return res.status(404).json({ message: "Tarefa não encontrada" });
+    } else {
+      return res.status(200).json(list);
+    }
+  } catch (err) {
+    return res.status(500).json({ error: err.message });
+  }
+});
+
 router.post("/listTarefs", async (req, res) => {
   try {
     const list = req.body;
@@ -30,9 +52,16 @@ router.put("/listTarefs/:id", async (req, res) => {
   const updatedData = req.body;
 
   try {
-    const updatedList = await listTarefs.findByIdAndUpdate(id, updatedData, {
-      new: true,
-    });
+    const updatedList = await listTarefs.findByIdAndUpdate(
+      {
+        _id: id,
+        userId: updatedData.userId, 
+      },
+      updatedData,
+      {
+        new: true,
+      }
+    );
     if (!updatedList) {
       return res.status(404).json({ message: "Tarefa não encontrada" });
     }
@@ -42,19 +71,19 @@ router.put("/listTarefs/:id", async (req, res) => {
   }
 });
 
-router.delete('/listTarefs/:id', async (req, res) => {
-    const { id } = req.params
+router.delete("/listTarefs/:id", async (req, res) => {
+  const { id } = req.params;
+  const { userId } = req.body;
 
-    try{
-        const deleteList = await listTarefs.findByIdAndDelete(id)
-        if(!deleteList){
-            return res.status(404).json({ message: 'Tarefa não encontrada' })
-        }
-        return res.status(200).json({ message: 'Tarefa deletada com sucesso ' })
+  try {
+    const deleteList = await listTarefs.findByIdAndDelete({ _id: id, userId});
+    if (!deleteList) {
+      return res.status(404).json({ message: "Tarefa não encontrada" });
     }
-    catch(err){
-        return res.status(400).json({ message: err.message })
-    }
-})
+    return res.status(200).json({ message: "Tarefa deletada com sucesso " });
+  } catch (err) {
+    return res.status(400).json({ message: err.message });
+  }
+});
 
-export default router
+export default router;
